Move custom type parser tests to its module
I see no reason why those should be in value_tests.
